[QNN EP] Extract neuron activations into a standalone op builder
The neuron activations (Elu, Gelu, Relu, Sigmoid, Tanh, HardSigmoid,
HardSwish, Softplus) all lower to QNN_OP_ELEMENT_WISE_NEURON but carried
per-op param construction, a Softplus rank guard, a Tanh uint16 QNN
validation workaround, and 16-bit output quant-param overrides inside
SimpleOpBuilder — again contradicting its "simple" contract. Move them
into a new NeuronOpBuilder.

Pure code move, no behavior change:
- New neuron_op_builder.cc with NeuronOpBuilder for the 8 neuron ops,
  including the moved ProcessNodeAttribute helper, OverrideQuantParams,
  and the OverrideOutputQuantParam override.
- Removed those blocks + the Softplus check + the dead
  ProcessSigmoidOrTanhOutput declaration from simple_op_builder.cc.
- Removed the stale CreateHardSigmoidOpBuilder declaration (never
  defined or called) from the factory header.
- Registered the 8 ops via CreateNeuronOpBuilder in the factory.
