+++ onnxruntime/core/providers/qnn/builder/op_builder_factory.cc
@@ -84,16 +84,16 @@ OpBuilderRegistrations::OpBuilderRegistrations() {
   CreateSimpleOpBuilder("DequantizeLinear", *this);
   CreateSimpleOpBuilder("DepthToSpace", *this);
   CreateSimpleOpBuilder("Div", *this);
-  CreateSimpleOpBuilder("Elu", *this);
+  CreateNeuronOpBuilder("Elu", *this);
   CreateSimpleOpBuilder("Equal", *this);
   CreateSimpleOpBuilder("Exp", *this);
   CreateSimpleOpBuilder("Floor", *this);
-  CreateSimpleOpBuilder("Gelu", *this);
+  CreateNeuronOpBuilder("Gelu", *this);
   CreateSimpleOpBuilder("Greater", *this);
   CreateSimpleOpBuilder("GreaterOrEqual", *this);
   CreateSimpleOpBuilder("GridSample", *this);
-  CreateSimpleOpBuilder("HardSigmoid", *this);
-  CreateSimpleOpBuilder("HardSwish", *this);
+  CreateNeuronOpBuilder("HardSigmoid", *this);
+  CreateNeuronOpBuilder("HardSwish", *this);
   CreateSimpleOpBuilder("LeakyRelu", *this);
   CreateSimpleOpBuilder("Less", *this);
   CreateSimpleOpBuilder("LessOrEqual", *this);
@@ -108,21 +108,21 @@ OpBuilderRegistrations::OpBuilderRegistrations() {
   CreateSimpleOpBuilder("Pow", *this);
   CreateSimpleOpBuilder("PRelu", *this);
   CreateSimpleOpBuilder("QuantizeLinear", *this);
-  CreateSimpleOpBuilder("Relu", *this);
+  CreateNeuronOpBuilder("Relu", *this);
   CreateSimpleOpBuilder("Round", *this);
   CreateScatterElementsOpBuilder("ScatterElements", *this);
   CreateScatterNDOpBuilder("ScatterND", *this);
   CreateSeluOpBuilder("Selu", *this);
   CreateShapeOpBuilder("Shape", *this);
-  CreateSimpleOpBuilder("Sigmoid", *this);
+  CreateNeuronOpBuilder("Sigmoid", *this);
   CreateSimpleOpBuilder("Sign", *this);
   CreateSimpleOpBuilder("Sin", *this);
-  CreateSimpleOpBuilder("Softplus", *this);
+  CreateNeuronOpBuilder("Softplus", *this);
   CreateSimpleOpBuilder("SpaceToDepth", *this);
   CreateSimpleOpBuilder("Sqrt", *this);
   CreateSimpleOpBuilder("Sub", *this);
   CreateSimpleOpBuilder("Sum", *this);
-  CreateSimpleOpBuilder("Tanh", *this);
+  CreateNeuronOpBuilder("Tanh", *this);
   CreateSimpleOpBuilder("Where", *this);
   CreateSimpleOpBuilder("Xor", *this);
   CreateSliceOpBuilder("Slice", *this);
